Post-verification adjustments!
