+++ src/main/java/com/Base.java
@@ -13,5 +13,4 @@ public class Base {
                 .setBaseUri(BASE_URL)
                 .build();
     }
-
 }
