+++ src/main/java/com/UserOperations.java
@@ -76,5 +76,4 @@ public class UserOperations {
                 .then()
                 .statusCode(202);
     }
-
 }
