+++ src/main/java/com/page/AuthPage.java
@@ -16,6 +16,18 @@ public class AuthPage {
     @FindBy(how = How.XPATH,using = ".//input[@name='name']")
     private SelenideElement emailInput;
 
+    @FindBy(how = How.XPATH,using = ".//input[@name='Пароль']")
+    private SelenideElement passwordInput;
+
+    @FindBy(how = How.XPATH,using = ".//button[text()='Войти']")
+    private SelenideElement enterButton;
+
+    @FindBy(how = How.XPATH,using = ".//a[text()='Зарегистрироваться']")
+    private SelenideElement registerLink;
+
+    @FindBy(how = How.XPATH,using = ".//a[text()='Восстановить пароль']")
+    private SelenideElement restorePasswordLink;
+
     @Step("Ввод логина")
     public AuthPage setEmail(String email) {
         emailInput.shouldBe(empty).click();
@@ -23,18 +35,12 @@ public class AuthPage {
         return this;
     }
 
-    @FindBy(how = How.XPATH,using = ".//input[@name='Пароль']")
-    private SelenideElement passwordInput;
-
     @Step("Ввод пароля")
     public AuthPage setPassword(String password) {
         passwordInput.shouldBe(empty).click();
         passwordInput.setValue(password);
         return this;
     }
-
-    @FindBy(how = How.XPATH,using = ".//button[text()='Войти']")
-    private SelenideElement enterButton;
 
     @Step("Нажать на кнопку Войти")
     public MainPage enterButtonClick() {
@@ -47,22 +53,15 @@ public class AuthPage {
         return enterButton.shouldBe(visible).isDisplayed();
     }
 
-    @FindBy(how = How.XPATH,using = ".//a[text()='Зарегистрироваться']")
-    private SelenideElement registerLink;
-
     @Step("Перейти по ссылке 'Зарегистрироваться'")
     public RegPage registerLinkClick() {
         registerLink.click();
         return page(RegPage.class);
     }
 
-    @FindBy(how = How.XPATH,using = ".//a[text()='Восстановить пароль']")
-    private SelenideElement restorePasswordLink;
-
     @Step("Перейти по ссылке 'Восстановить пароль'")
     public RestorePassPage restorePasswordLinkClick() {
         restorePasswordLink.click();
         return page(RestorePassPage.class);
     }
-
 }
