+++ src/main/java/com/page/MainPage.java
@@ -15,14 +15,44 @@ public class MainPage {
     @FindBy(how = How.XPATH, using = ".//button [text()='Войти в аккаунт']")
     public SelenideElement signInButton;
 
+    @FindBy(how = How.XPATH,using = "//p[text()='Личный Кабинет']")
+    public SelenideElement profileButton;
+
+    @FindBy (how = How.XPATH, using = "//span[text()='Булки']")
+    private SelenideElement bunsTab;
+
+    @FindBy (how = How.XPATH, using = "//span[text()='Соусы']")
+    private SelenideElement saucesTab;
+
+    @FindBy (how = How.XPATH, using = "//span[text()='Начинки']")
+    private SelenideElement fillingsTab;
+
+    @FindBy(how = How.XPATH,using = "//h2[text()='Булки']") ////h2[text()='Булки']
+    private SelenideElement bunsTabSelected;
+
+    @FindBy(how = How.XPATH,using = "//h2[text()='Начинки']")//h2[text()='Начинки']
+    private SelenideElement fillingsTabSelected;
+
+    @FindBy(how = How.XPATH,using = "//h2[text()='Соусы']")//h2[text()='Соусы']
+    private SelenideElement saucesTabSelected;
+
+    @FindBy(how = How.XPATH,using = "//p[text()='Флюоресцентная булка R2-D3']")
+    private SelenideElement bunFlu;
+
+    @FindBy(how = How.XPATH,using = "//p[text()='Соус Spicy-X']")
+    private SelenideElement sauceSpicy;
+
+    @FindBy(how = How.XPATH,using = "//p[text()='Мясо бессмертных моллюсков Protostomia']")
+    private SelenideElement fillingMeat;
+
+    @FindBy (how = How.XPATH, using = ".//button[text()='Оформить заказ']")
+    private SelenideElement executeOrderButton;
+
     // вход по кнопке «Войти в аккаунт» на главной
     public AuthPage signInButtonClick (){
         signInButton.click();
         return page(AuthPage.class);
     }
-
-    @FindBy(how = How.XPATH,using = "//p[text()='Личный Кабинет']")
-    public SelenideElement profileButton;
 
     //вход через кнопку «Личный кабинет»
     public AuthPage profileButtonClick (){
@@ -36,17 +66,11 @@ public class MainPage {
         return page(ProfilePage.class);
     }
 
-    @FindBy (how = How.XPATH, using = ".//span [@class='text text_type_main-default'][text()='Булки']")
-    private SelenideElement bunsTab;
-
     @Step("Нажать «Булки»")
     public MainPage bunsTabClick(){
         bunsTab.shouldBe(enabled).click();
         return this;
     }
-
-    @FindBy (how = How.XPATH, using = ".//span [@class='text text_type_main-default'][text()='Соусы']")
-    private SelenideElement saucesTab;
 
     @Step ("Нажать «Соусы»")
     public MainPage saucesTabClick(){
@@ -54,68 +78,43 @@ public class MainPage {
         return this;
     }
 
-    @FindBy (how = How.XPATH, using = ".//span [@class='text text_type_main-default'][text()='Начинки']")
-    private SelenideElement fillingsTab;
-
     @Step ("Нажать «Начинки»")
     public MainPage fillingsTabClick(){
         fillingsTab.shouldBe(enabled).click();
         return this;
     }
 
-    @FindBy(how = How.XPATH,using = "//div[contains(@class, 'current')]/span[text()='Булки']")
-    private SelenideElement bunsTabSelected;
-
     @Step("Проверить отображение выбранной вкладки «Булки».")
-    public boolean isBunTabSelected() {
+    public boolean isBunTabIsDisplayed() {
       return bunsTabSelected.shouldBe(visible).isDisplayed();
     }
 
-    @FindBy(how = How.XPATH,using = "//div[contains(@class, 'current')]/span[text()='Начинки']")
-    private SelenideElement fillingsTabSelected;
-
     @Step("Проверить отображение выбранной вкладки «Начинки».")
-    public boolean isFillingsTabSelected() {
+    public boolean isFillingsTabIsDisplayed() {
         return fillingsTabSelected.shouldBe(visible).isDisplayed();
     }
 
-    @FindBy(how = How.XPATH,using = "//div[contains(@class, 'current')]/span[text()='Соусы']")
-    private SelenideElement saucesTabSelected;
-
     @Step("Проверить отображение выбранной вкладки «Соусы».")
-    public boolean isSaucesTabSelected() {
+    public boolean isSaucesTabIsDisplayed() {
         return saucesTabSelected.shouldBe(visible).isDisplayed();
     }
-
-    @FindBy(how = How.XPATH,using = "//p[text()='Флюоресцентная булка R2-D3']")
-    private SelenideElement bunFlu;
 
     @Step("Проверить отображение булки «Флюоресцентная булка R2-D3»")
     public boolean isBunFluVisible() {
      return bunFlu.shouldBe(visible).isDisplayed();
     }
 
-    @FindBy(how = How.XPATH,using = "//p[text()='Соус Spicy-X']")
-    private SelenideElement sauceSpicy;
-
     @Step("Проверить отображение соуса «Соус Spicy-X»")
     public boolean isSauceSpicyVisible() {
        return sauceSpicy.shouldBe(visible).isDisplayed();
     }
-
-    @FindBy(how = How.XPATH,using = "//p[text()='Мясо бессмертных моллюсков Protostomia']")
-    private SelenideElement fillingMeat;
 
     @Step("Проверить отображение начинки «Мясо бессмертных моллюсков Protostomia»")
     public boolean isFillingMeatVisible() {
         return fillingMeat.shouldBe(visible).isDisplayed();
     }
 
-    @FindBy (how = How.XPATH, using = ".//button[text()='Оформить заказ']")
-    private SelenideElement executeOrderButton;
-
     public boolean executeOrderButtonVisible (){
         return executeOrderButton.shouldBe(visible).isDisplayed();
     }
-
 }
