+++ src/main/java/com/page/ProfilePage.java
@@ -14,15 +14,24 @@ public class ProfilePage {
     @FindBy(how = How.XPATH,using = ".//button[text()='Выход']")
     private SelenideElement logoutButton;
 
+    @FindBy (how = How.XPATH, using = "//*[text()='Конструктор']")
+    private SelenideElement constructorButton;
+
+    @FindBy (how = How.XPATH, using = "//*[@class='AppHeader_header__logo__2D0X2']")
+    private SelenideElement burgerLogo;
+
+    @FindBy(how = How.XPATH,using = "//input[@name=\"Name\"]")
+    private SelenideElement nameInput;
+
+    @FindBy(how = How.XPATH,using = "//input[@type=\"text\" and @name=\"name\"]")
+    private SelenideElement loginInput;
+
     //выход по кнопке «Выйти» в личном кабинете
     @Step("Нажать на кнопку 'Выход'")
     public MainPage exitButtonClick(){
         logoutButton.click();
         return page(MainPage.class);
     }
-
-    @FindBy (how = How.XPATH, using = "//*[text()='Конструктор']")
-    private SelenideElement constructorButton;
 
     //переход по клику на «Конструктор»
     @Step("Нажать на кнопку 'Конструктор'")
@@ -31,9 +40,6 @@ public class ProfilePage {
         return page(MainPage.class);
     }
 
-    @FindBy (how = How.XPATH, using = "//*[@class='AppHeader_header__logo__2D0X2']")
-    private SelenideElement burgerLogo;
-
     //Переход из личного кабинета в конструктор
     @Step("Нажать на лого")
     public MainPage burgerLogoClick() {
@@ -41,18 +47,11 @@ public class ProfilePage {
         return page(MainPage.class);
     }
 
-    @FindBy(how = How.XPATH,using = "(//div[@class='input__container']//div/input[@type='text'])[1]")
-    private SelenideElement nameInput;
-
     public String getNameInputValue () {
         return nameInput.getValue();
     }
 
-    @FindBy(how = How.XPATH,using = "(//div[@class='input__container']//div/input[@type='text'])[2]")
-    private SelenideElement loginInput;
-
     public String getLoginInputValue () {
         return loginInput.getValue();
     }
-
 }
