+++ src/test/java/AuthPageTest.java
@@ -1,33 +1,15 @@
 import com.page.*;
-import com.UserOperations;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
 import static com.codeborne.selenide.Selenide.*;
-import static com.page.MainPage.MAIN_PAGE_URL;
 import static org.junit.Assert.assertTrue;
-import java.util.Map;
 
 public class AuthPageTest extends Config {
-    UserOperations userOperations = new UserOperations();
-    Map<String, String> newUser = userOperations.register();
+
     String email = newUser.get("email");
     String password = newUser.get("password");
-    MainPage mainPage;
-
-    @Before
-    public void before() {
-        mainPage = open(MAIN_PAGE_URL, MainPage.class);
-    }
-
-    @After
-    public void tearDown() {
-        userOperations.delete();
-        webdriver().driver().close();
-    }
 
     @Test
     @DisplayName("Проверка, что пользователь может войти по кнопке «Войти в аккаунт» на Главной")
@@ -73,7 +55,6 @@ public class AuthPageTest extends Config {
     @Test
     @DisplayName("Проверка, что пользователь может войти по ссылке «Войти» на форме восстановления пароля")
     @Description("При успешном входе на Главной появится кнопка «Оформить заказ»")
-
     public void loginUserWithSignInLinkOnRestorePageTest(){
         mainPage
                 .signInButtonClick()
@@ -85,5 +66,4 @@ public class AuthPageTest extends Config {
         boolean isExecuteOrderButtonDisplayedNow = page(MainPage.class).executeOrderButtonVisible();
         assertTrue("Кнопка «Оформить заказ» не появилась", isExecuteOrderButtonDisplayedNow);
     }
-
 }
