+++ src/test/java/ConstructorTest.java
@@ -1,26 +1,10 @@
-import com.page.MainPage;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
-import static com.codeborne.selenide.Selenide.*;
-import static com.page.MainPage.MAIN_PAGE_URL;
 import static org.junit.Assert.assertTrue;
 
 public class ConstructorTest extends Config {
-    MainPage mainPage;
-
-    @Before
-    public void before() {
-        mainPage = open(MAIN_PAGE_URL, MainPage.class);
-    }
-
-    @After
-    public void tearDown() {
-        webdriver().driver().close();
-    }
 
     @Test
     @DisplayName("Проверка, что пользователь может перейти на вкладку «Соусы» на Главной")
@@ -29,7 +13,7 @@ public class ConstructorTest extends Config {
         boolean isSaucesSectionDisplayed =
                 mainPage
                         .saucesTabClick()
-                        .isSaucesTabSelected();
+                        .isSaucesTabIsDisplayed();
         assertTrue("Секция «Соусы» не появилась", isSaucesSectionDisplayed);
         assertTrue("Соус Spicy-X не появился", mainPage.isSauceSpicyVisible());
     }
@@ -41,8 +25,9 @@ public class ConstructorTest extends Config {
         boolean isBunsSectionDisplayed =
                 mainPage
                         .saucesTabClick()
+                        .fillingsTabClick()
                         .bunsTabClick()
-                        .isBunTabSelected();
+                        .isBunTabIsDisplayed();
         assertTrue("Секция «Булки» не появилась", isBunsSectionDisplayed);
         assertTrue("Флюоресцентная булка R2-D3 не появилась", mainPage.isBunFluVisible());
     }
@@ -54,7 +39,7 @@ public class ConstructorTest extends Config {
         boolean isFillingsSectionDisplayed =
                 mainPage
                         .fillingsTabClick()
-                        .isFillingsTabSelected();
+                        .isFillingsTabIsDisplayed();
         assertTrue("Секция «Начинки» не появилась", isFillingsSectionDisplayed);
         assertTrue("Мясо бессмертных моллюсков Protostomia не появилось", mainPage.isFillingMeatVisible());
     }
