+++ src/test/java/RegPageTest.java
@@ -1,41 +1,20 @@
-import com.UserOperations;
 import com.page.*;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import java.time.Instant;
-import java.util.Map;
 
 import static com.codeborne.selenide.Selenide.*;
-import static com.page.MainPage.MAIN_PAGE_URL;
 import static org.junit.Assert.*;
 
 public class RegPageTest extends Config {
-    UserOperations userOperations = new UserOperations();
-    Map<String, String> newUser = userOperations.register();
+
     String email = newUser.get("email");
     String password = newUser.get("password");
     String name = newUser.get("name");
-    MainPage mainPage;
     RegPage regPage = page(RegPage.class);
-
-
     Instant instant = Instant.now();
     long timeStampMillis = instant.toEpochMilli(); //префикс для сгенерированных email
-
-    @Before
-    public void before() {
-        mainPage = open(MAIN_PAGE_URL, MainPage.class);
-
-    }
-
-    @After
-    public void tearDown() {
-        userOperations.delete();
-        webdriver().driver().close();
-    }
 
     @Test
     @DisplayName("Проверка, что пользователь может зарегистрироваться и затем авторизоваться")
